Reorganizing to keep project readable and tidy. Implemented own ModelErrors that can be easily mapped to other errors.
