+++ backend/src/routes/create_questions.rs
@@ -3,70 +3,21 @@
 use crate::{error_chain_helper, session_wrapper::SessionWrapper, surrealdb_repo::Database};
 use actix_web::http::{header::ContentType, StatusCode};
 use actix_web::{web, HttpRequest, HttpResponse, ResponseError};
-use models::questions::{QuestionMC, SurrealQuestionMC};
-use models::SurrealRecord;
+use models::{
+    model_errors::ModelErrors,
+    questions::{JsonQuestion, JsonQuestionMC, QuestionJsonPkg, QuestionMC, SurrealQuestionMC},
+    SurrealRecord,
+};
 use serde::{Deserialize, Serialize};
 use surrealdb::opt::PatchOp;
 use surrealdb::sql::{Id, Thing};
 use uuid::Uuid;
 
-#[derive(Serialize, Deserialize, Debug)]
-pub struct JsonPkg {
-    pub quiz_id: Thing,
-    pub questions: Vec<JsonQuestion>,
-}
-
-/// To allow for the easy transporation of data
-/// If adding another type, be sure to update the `JsonPkg::validate_fields()` method.
-#[derive(Serialize, Deserialize, Debug)]
-pub enum JsonQuestion {
-    MultipleChoice(JsonQuestionMC),
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct JsonQuestionMC {
-    pub question: String,
-    pub hint: Option<String>,
-    pub answer: String,
-    pub choices: Vec<String>,
-}
-
-impl JsonPkg {
-    fn validate_fields(&self) -> Result<(), CreateQuestionError> {
-        // The type system ensures quiz_id isn't empty
-        use JsonQuestion::*;
-        for (num, what) in self.questions.iter().enumerate() {
-            match what {
-                MultipleChoice(qmc) => {
-                    if qmc.question.trim().len() < 1 {
-                        return Err(CreateQuestionError::ValidationError(format!(
-                            "Q{}: Question cannot be empty",
-                            num
-                        )));
-                    } else if qmc.answer.trim().len() < 1 {
-                        return Err(CreateQuestionError::ValidationError(format!(
-                            "Q{}: Question needs valid answer",
-                            num
-                        )));
-                    } else if qmc.choices.len() < 1 {
-                        return Err(CreateQuestionError::ValidationError(format!(
-                            "Q{}: Question needs at least one additional choice",
-                            num
-                        )));
-                    }
-                    // Could loop through choices to ensure they are also not blank
-                }
-            }
-        }
-        Ok(())
-    }
-}
-
 // -- Errors --
 #[derive(thiserror::Error)]
 pub enum CreateQuestionError {
-    #[error("{0}")]
-    ValidationError(String),
+    #[error(transparent)]
+    ValidationError(#[from] ModelErrors),
     #[error("{0}")]
     AuthorizationError(String),
     #[error(transparent)]
@@ -112,9 +63,9 @@ pub async fn create_new_questions(
     req: HttpRequest,
     session: SessionWrapper,
     db: web::Data<Database>,
-    question_pkg_pt: web::Json<JsonPkg>,
+    question_pkg_pt: web::Json<QuestionJsonPkg>,
 ) -> Result<HttpResponse, CreateQuestionError> {
-    let question_data: JsonPkg = question_pkg_pt.into_inner();
+    let question_data: QuestionJsonPkg = question_pkg_pt.into_inner();
     question_data.validate_fields()?;
 
     let quiz_id: Thing = question_data.quiz_id;
