+++ backend/src/routes/create_quiz.rs
@@ -2,61 +2,18 @@
 use crate::{error_chain_helper, session_wrapper::SessionWrapper, surrealdb_repo::Database};
 use actix_web::http::{header::ContentType, StatusCode};
 use actix_web::{web, HttpRequest, HttpResponse, ResponseError};
-use serde::{Deserialize, Serialize};
-use surrealdb::sql::{Id, Thing};
+use models::{
+    model_errors::ModelErrors,
+    quiz::{Quiz, QuizJsonPkg, SurrealQuiz},
+};
+use surrealdb::sql::Id;
 use uuid::Uuid;
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct JsonPkg {
-    name: String,
-    description: String,
-}
-
-impl JsonPkg {
-    fn validate_field(&self) -> Result<(), CreateQuizError> {
-        if self.name.trim().len() < 1 {
-            Err(CreateQuizError::ValidationError(String::from(
-                "Quiz name cannot be blank or white space",
-            )))
-        } else {
-            Ok(())
-        }
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct FullQuiz {
-    pub id: Thing,
-    pub name: String,
-    pub description: String,
-    pub author_id: String,
-    pub questions_mc: Vec<Thing>,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct Quiz {
-    pub name: String,
-    pub description: String,
-    pub author_id: String,
-    pub questions_mc: Vec<Thing>,
-}
-
-impl Quiz {
-    pub fn new(name: String, description: String, author_id: String) -> Self {
-        Self {
-            name,
-            description,
-            author_id,
-            questions_mc: Vec::new(),
-        }
-    }
-}
 
 // Errors
 #[derive(thiserror::Error)]
 pub enum CreateQuizError {
-    #[error("{0}")]
-    ValidationError(String),
+    #[error(transparent)]
+    ValidationError(#[from] ModelErrors),
     #[error("{0}")]
     AuthorizationError(String),
     #[error(transparent)]
@@ -78,9 +35,9 @@ impl ResponseError for CreateQuizError {
                     .insert_header(ContentType::json())
                     .json(serde_json::json!({"msg": "Unknown Error"}))
             }
-            CreateQuizError::ValidationError(msg) => HttpResponse::build(StatusCode::BAD_REQUEST)
+            CreateQuizError::ValidationError(err) => HttpResponse::build(StatusCode::BAD_REQUEST)
                 .insert_header(ContentType::json())
-                .json(serde_json::json!({ "msg": msg })),
+                .json(serde_json::json!({ "msg": err })),
             CreateQuizError::AuthorizationError(msg) => {
                 HttpResponse::build(StatusCode::UNAUTHORIZED)
                     .insert_header(ContentType::json())
@@ -99,9 +56,9 @@ pub async fn create_new_quiz(
     req: HttpRequest,
     session: SessionWrapper,
     db: web::Data<Database>,
-    quiz_pkg_pt: web::Json<JsonPkg>,
+    quiz_pkg_pt: web::Json<QuizJsonPkg>,
 ) -> Result<HttpResponse, CreateQuizError> {
-    let quiz_data: JsonPkg = quiz_pkg_pt.into_inner();
+    let quiz_data: QuizJsonPkg = quiz_pkg_pt.into_inner();
     quiz_data.validate_field()?;
 
     let some_user_id: Option<Uuid> = session
@@ -122,12 +79,12 @@ pub async fn create_new_quiz(
     dbg!(&quiz_to_save);
     dbg!(Id::uuid().to_string());
 
-    let created: Option<FullQuiz> = db
+    let created: Option<SurrealQuiz> = db
         .client
         .create(("quizzes", Id::uuid().to_string()))
         .content(&quiz_to_save)
         .await
-        .map_err(|e| CreateQuizError::ValidationError(e.to_string()))?;
+        .map_err(|e| CreateQuizError::UnexpectedError(anyhow::anyhow!(e)))?;
 
     // Dummy response to shut up linter
     Ok(HttpResponse::Ok().json(&created))
